Move bindings and binding traits to bindings mod
