+++ src/bindings/mod.rs
@@ -1,0 +1,3 @@
+pub mod in_memory_index;
+pub mod memmap_index;
+pub mod sharded_memmap_index;
