+++ src/lib.rs
@@ -1,20 +1,18 @@
 pub mod mmap_slice;
-pub use in_memory_index::InMemoryIndex;
-pub use memmap_index::MemmapIndex;
-pub use sharded_memmap_index::ShardedMemmapIndex;
+pub use bindings::in_memory_index::InMemoryIndex;
+pub use bindings::memmap_index::MemmapIndex;
+pub use bindings::sharded_memmap_index::ShardedMemmapIndex;
 pub use table::SuffixTable;
 
 /// Python bindings
 use pyo3::prelude::*;
 
+mod bindings;
 mod in_memory_index;
-mod in_memory_index_rs;
 mod memmap_index;
-mod memmap_index_rs;
 mod par_quicksort;
 mod sample;
 mod sharded_memmap_index;
-mod sharded_memmap_index_rs;
 mod table;
 mod util;
